Move getApiUrl to a new utils folder
