+++ .env
@@ -1,3 +1,3 @@
 # If you are altering the dev URL, remember to update the port in backend/src/Properties/launchSettings.json
-API_URL_DEVELOPMENT=http://localhost:8080
-API_URL_PRODUCTION=https://localhost:8081
+API_URL_DEVELOPMENT=http://localhost:8080/api
+API_URL_PRODUCTION=https://localhost:8081/api
