+++ backend/src/Controllers/SubmissionsController.cs
@@ -1,7 +1,6 @@
 using FormAPI.Mappers;
 using FormAPI.Models;
 using FormAPI.Repositories;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FormAPI.Controllers;
