+++ frontend/main-app/src/components/DashboardOverview/DashboardOverview.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { getApiUrl } from "../Login/LoginUtils";
+import { getApiUrl } from "../../utils/getApiUrl";
 import axios from "axios";
 import classes from "./DashboardOverview.module.css";
 import { Heading, Paragraph } from "@digdir/design-system-react";
@@ -16,7 +16,7 @@ export const DashboardOverview: React.FC<DashboardOverviewProps> = ({ forms }) =
   useEffect(() => {
     (async function getSubmissionCount() {
       const apiUrl = getApiUrl();
-      const targetUrl = `${apiUrl}/api/submissions/`;
+      const targetUrl = `${apiUrl}/submissions/`;
 
       try {
         const result = await axios.get(targetUrl);
