+++ frontend/main-app/src/components/FormViewer/httpUtils.tsx
@@ -1,9 +1,9 @@
-import { getApiUrl } from "../Login/LoginUtils";
+import { getApiUrl } from "../../utils/getApiUrl";
 import axios from "axios";
 
 export const getFormSchema = async (formId: string, setFormSchema) => {
   const apiUrl = getApiUrl();
-  const targetUrl = `${apiUrl}/api/forms/${formId}`;
+  const targetUrl = `${apiUrl}/forms/${formId}`;
 
   try {
     const response = await axios.get(targetUrl);
@@ -31,7 +31,7 @@ const formatSubmission = (formSchema, formData) => {
 export const postSubmission = async (formSchema, formData, setFormAlert) => {
   const formattedSubmission = formatSubmission(formSchema, formData);
   const apiUrl = getApiUrl();
-  const targetUrl = `${apiUrl}/api/submissions/`;
+  const targetUrl = `${apiUrl}/submissions/`;
 
   try {
     await axios.post(targetUrl, formattedSubmission);
