+++ frontend/main-app/src/components/UserDropdown/UserDropdown.module.css
@@ -1,3 +1,3 @@
 .userIcon {
-    color: var(--fds-semantic-text-neutral-default);
+  color: var(--fds-semantic-text-neutral-default);
 }
